# PERF: speed-up when scalar not found in Categorical's categories