@jreback 2019-11-21
thanks @topper-123 any additional asv's for this / categorical stuff surely welcome.